move templates as package
